backend/services/witsml/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
paho-mqtt==2.0.0
requests==2.31.0
